demo/python/processor/requirements.txt
requests>=2.32.0,<3.0.0
opentelemetry-instrumentation-requests>=0.52b1,<1.0.0
opentelemetry-api>=1.31.1,<2.0.0
lambda-otel-lite>=0.11.0,<1.0.0
opentelemetry-instrumentation-boto3sqs>=0.52b1,<1.0.0
opentelemetry-propagator-aws-xray>=1.0.2,<2.0.0

